Remove thread-local execution spaces
Instead add a helper struct that can be initialized for a varying number of
worker threads and with a varying number of instance per thread. This ensures
that the instances can be destructed before Kokkos::finalize. This also removes
the compiled library options as they are not needed anymore; this is once again
header-only.